queryDsl 변환 작업 완료

+++ banchanshop/src/main/java/com/bbsk/banchanshop/admin/dto/ResponseRecently6MonthsOrderCount.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class ResponseRecently6MonthsOrderCount {
 
     private String month;
-    private int orderCount;
+    private Long orderCount;
 }
